Adding newton_refine() for curve-curve intersections.
In the process, cleaned up code that created matrices for
simple combinations of rows that could accomplished via
linear combinations of slices.

Also factored out the behavior of Curve.evaluate()
into a helper since the same helper is useful when
evaluating the Hodograph curve (which is needed for
a Newton step).