Adding handle_corners() helper.
This takes an edge-edge intersection (from within a surface)
and moves any corner intersection from the end of an edge
to the beginning of the next edge.

This is to avoid duplication and allow easier classification
of intersects at corners.

As a result, tweaking classify_intersection() to **fail**
on the end of segments and treat the beginning of segments
like normal (no more -2 returned has some other side
effects too).

Also updating the corner-classify example to be clearer.

Also updating surface-surface functional tests based on
the correct diagnosis of corner intersections.