Using less exact measure of error in curve-curve-intersection.
We really only want the size of the error, so the exponent
is all that is required anyhow.

This change was "required" because I bought a new laptop.
The previous max error in ``intersect-8-27`` was

>>> np.log2(10./(2**54))
-50.678071905112638

Due to a difference of a single bit in the computed solution,
the error became

>>> np.log2(11./(2**54))
-50.540568381362704

What caused this one bit difference between two computers
running identical software? BLAS operations tailored to
a particular machine.

In this case, the error came when calling curve2.evaluate(t)
on CURVE27, with

>>> t = float.fromhex('0x1.b9696fbb6feeap-2')
>>> t
0.43106627065206615

To evaluate the point on the curve, the Bernstein basis
functions are computed at t and give

>>> weights = np.array([[
...     float.fromhex('0x1.792698cc8668ep-3'),
...     float.fromhex('0x1.aca2cf310eea7p-2'),
...     float.fromhex('0x1.44c42d38c8c1ep-2'),
...     float.fromhex('0x1.4816dcbf947cdp-4'),
... ]])
>>> weights
array([[ 0.18415565,  0.41858982,  0.3171546 ,  0.08009993]])

This is true on all platforms: they are just computed using
standard flops on IEEE-754 doubles.

The right hand side is also fixed (constant) on both platforms,
it is just the nodes in CURVE27:

>>> nodes = np.array([
...     [0.125, 0.25  ],
...     [0.375, 0.75  ],
...     [0.625, 0.0   ],
...     [0.875, 0.1875],
... ])

But when computing the (1 x 4) times (4 x 2) matrix product,
there is a one-bit difference. One my old laptop, the
value produced was equal to that via a serial addition
from the first row to the end:

>>> res_serial = (
...     weights[0, 0] * nodes[0, :] +
...     weights[0, 1] * nodes[1, :] +
...     weights[0, 2] * nodes[2, :] +
...     weights[0, 3] * nodes[3, :]
... )
>>> res_serial
array([ 0.4482997 ,  0.37500001])
>>> res_serial[0].hex()
'0x1.cb0f13cc93f2fp-2'
>>> res_serial[1].hex()
'0x1.800000d756f2dp-2'

On the new laptop, it was instead equal to the value produced
via a blocked addition: add each pair of rows, then add the
two values produced:

>>> res_blocked = (
...     (
...         weights[0, 0] * nodes[0, :] +
...         weights[0, 1] * nodes[1, :]
...     ) + (
...         weights[0, 2] * nodes[2, :] +
...         weights[0, 3] * nodes[3, :]
...     )
... )
>>> res_blocked
array([ 0.4482997 ,  0.37500001])
>>> res_blocked[0].hex()
'0x1.cb0f13cc93f30p-2'
>>> res_blocked[1].hex()
'0x1.800000d756f2dp-2'

The resulting values only differed in the first component and
the different was in the very last bit:

>>> 0xcb0f13cc93f30 - 0xcb0f13cc93f2f
1