Fixing round-off issues (between platforms).
By inspecting (on a local Windows install), the
newton-refine-curve-cusp-continued doctest went from 27 to
39 values because the 5th computed s-value differed on
Windows in the final bit. This is similar to the
previous change, where BLAS operations are slightly
different across platforms (and here, newton_refine has
**many** BLAS operations).

Also upgrading the number of wiggle bits in the 1Q and 3Q
functional test.