Removing some features from Base.
This is because they were a little "incorrect". The equality
check was not "future-proof" as subclasses add features and
the copy was not general enough to take into account new values
added.

This is in preparation for more streamlined use of shared
properties (i.e. passing an optional ``degree`` to ``_get_degree``
to avoid re-computing).