Doing some loop-unrolling in _tangent_bbox_intersection().
This may not actually be an optimization, e.g. if the function
call overhead outweighs the for-loop overhead (both of which
are nonzero in Python).

However, this was done not only to unroll the teeny-tiny
loops, but also to avoid creating each ``node_right`` value
twice (even though a view, still an extra object).