Taking corner values into account in polynomial_sign().
Fixes #12.

The test case added was overly specific, but it came from the
issue where I first realized ``is_valid`` could be improved
without much work.

Any coefficients of the form

    [X, *, 0.0, *, *, Y]

with ``X, Y > 0`` would have done the trick. It would be
nice to show that it fails to make a conclusion without
taking the signs at the corner nodes into consideration.
It would also be nice to have an example that didn't help
until subdivision.