Converting other NumPy matrix operations to use Fortran ordering.
This is a follow-up to using np.asfortranarray() over np.array().
In particular, making `np.zeros` and `np.emtpy` use Fortran-ordering
on creation and adding a custom version of `np.eye` that uses
Fortran ordering (and just uses the same implementation). Also
using Fortran ordering in `.copy()` and `.reshape()`.

Implemented some helpers for taking a matrix product as well.

Updating Curve subdivision so that the new nodes created are
contiguous. Previously, we were grabbing a slice of the last half
rows, so since Fortran/column contiguous, the sliced nodes
were no longer contiguous.