Adding Fortran implementation of `_wiggle_interval`.
This wasn't necessarily needed but it's a means to an end: adding
a Fortran implementation of `from_linearized()`.

In order to raise a Python error from the Fortran function (via
`f2py`), added a `.pyf` file. Needed to utilize the `callstatement`
feature of `f2py` for this.

H/T to:

https://mail.scipy.org/pipermail/numpy-discussion/2009-January/039672.html
https://gist.github.com/dhermes/81486f13dc30a48c5622981d3b87a093