Re-vamping "wiggle interval" to round **towards** 0 and 1.
This is in contrast to the previous approach, where we only rounded
to those values from outside of [0, 1]. But, a programming error
caused an infinite loop in ``_basic_interior_combine`` (see the
test cases added to fnl. tests) that revealed an error: almost
corners (e.g. t = 0.9999999999999971 = 1 - 26 * 2^{-53}) are not
corners.

As a result, also added a bail-out (failure) in ``_basic_interior_combine``
when unexpected behavior occurs. Also revamped the intersection
verification step to check for "almost" exactness rather than bitwise
identical duplicates.