Adding images from test_segment_box fnl. test.
This required a slight modification there and in
runtime_utils to allow for an extra piece to be added
to the filename.

In the process, `test_goes_through_box.png` was renamed to
`test_goes_through_box08.png` because there were 9 examples and
the final one is the only one that was actually saved. This
also forced an update of a reference to this image in
`DEVELOPMENT.rst`.