Adding polynomial information to curve_intersections.json.
These are the coefficients of the polynomials that the s- and t-
parameters are roots of. For example, in intersection #16, the
following were added:

  curve1_poly -> [1, -12, 16]
  curve2_poly -> [5, -20, 16]

This means that the roots of `1 - 12 s + 16 s^2` are among the
parameters in "curve1_params", and indeed we see that:

  "0x1.8722191a02d61p-4" ~= (3 - sqrt(5)) / 8
  "0x1.4f1bbcdcbfa54p-1" ~= (3 + sqrt(5)) / 8

and similarly with "curve2_params":

  "0x1.61c8864680b58p-2" ~= (5 - sqrt(5)) / 8
  "0x1.cf1bbcdcbfa54p-1" ~= (5 + sqrt(5)) / 8

In all cases, the polynomial coefficients

- are integers
- represent an irreducible polynomial
- come in "index<->exponent" order, i.e. index 0 is the constant
  term, index 1 is the linear term, index 2 is the quadratic
  term, etc.

For some of the intersection cases, there are rational **and**
irrational parameters, but the polynomial only applies to
the irrational parameters.

As it turns out, none of the curve-curve intersections have
irrational parameters that are roots of **different** irreducible
polynomials. If that were the case, then "curve1_poly" / "curve2_poly"
would need to be lists of lists or some other data structure that
could capture the mapping from parameter to polynomial.