Adding information about implicitized algebraic curves.
For the curves (other than 40 and 41) that have a "reasonable"
algebraic representation, we implicitize via the resultant.

For example, for "curve 2"

    x(s) = (9 - 8 s) / 8
    y(s) = (2 s - 1)^2 / 2

so we have

    8 s - 4 = 5 - 8 x
    32 y = (8 s - 4)^2 = (5 - 8 x)^2
    25 - 32 y - 80 x + 64 x^2 = 0

which corresponds to

    [
        [ 25, 0, 0],
        [-32, 0, 1],
        [-80, 1, 0],
        [ 64, 2, 0],
    ]

This representation is a list of triples of integers:

    (coeff, degree_x, degree_y)

where `degree_x` and `degree_y` are non-negative. Though
it's not required, we also have `coeff != 0` and the triples
are sorted by total degree (`degree_x + degree_y`) and then
by `degree_x` to break ties.