Moving most of Surface.subdivide() logic into helper.
This is the **start** of an effort to move all actions on
"high-level" objects into the public-facing modules `curve`,
`surface` and `curved_polygon`.

With this change, `locate_point` (and transitively `_mean_centroid`) in
`_surface_helpers` now deal with raw nodes rather than with a
`Surface`. (This may have revealed a bug when using `Surface.locate()`
with non-default `_base_x` and `_base_y`.)