Use a nasty hack to get doctests to pass on AppVeyor.
Failure:
https://ci.appveyor.com/project/dhermes/bezier/build/1.0.404.master/job/he37rn22qfuvj5pd

The issue is that (for now), the `.../site-packages/bezier/lib`
directory will not exist in a Windows install. Eventually (once
the install issue is resolved) the `.../bezier/lib` directory
will exist, but it will contain `bezier.lib` rather than
`libbezier.a`, so this hack will need to be modified.