Using "custom" `build_ext` in `setup.py`.
This way, we **only** build the Fortran source files
when an extension is being built (so for other commands,
e.g. `egg_info`, there is no wasted compilation).

Also tried to re-structure `setup.py` a bit so that the
file paths were less ad-hoc and were specified in a single
place.

Also made `distutils.ccompiler` a global import since it is
part of the standard library.