Add template for DEVELOPMENT.rst as well.
This way a release specific version can be made that
pins to a fixed docset and a fixed commit on GitHub.

Also, this may allow future tweaking so that
``docs/development.rst`` is a Sphinx-specific file
rather than a symlink. For example, the link
https://bezier.readthedocs.io/en/{rtd_version}/curve-curve-intersection.html
could be tweaked so that it was just :doc:`curve-curve-intersection`
and the
https://raw.githubusercontent.com/dhermes/bezier/{revision}/docs/images/curves11_and_26.png
links could become relative links within the Sphinx docs.