Adding speedup for `compute_length`.
In the process, had to add a subset of QUADPACK and `d1mach.f`
in order to do adaptive quadrature to compute the length.

F77 source was retrieved directly from:

  http://www.netlib.org/quadpack/dqpsrt.f
  http://www.netlib.org/quadpack/dqk21.f
  http://www.netlib.org/quadpack/dqelg.f
  http://www.netlib.org/quadpack/dqagse.f
  http://www.netlib.org/blas/d1mach.f

It's worth noting that all of QUADPACK (and I believe `d1mach`) are
"Public Domain".

Used `dqagse` because that is what is called by `scipy.integrate.quad`
when there are no `points` (where the function behaves badly), there
is no special weight function and the bounds are finite.

I also used the default (from SciPy) `limit` of 50 subintervals in
the adaptive quadrature and both an absolute (`epsabs`) and relative
(`epsrel`) tolerance of `2^{-26}` (i.e. `sqrt(machine precision`).

It's worth noting that `setup_helpers.py` is a little unwieldy now
because there is so much copy-pasting in the dependencies.

For now, the error code `ier` is ignored in the Fortran implementation.
As comparision, it'd be at the very least printed as a warning from
the SciPy implementation.