Adding speedup for `ulps_away`.
At this point, every "computational" function in
`_helpers` and `_curve_helpers` has been ported to Fortran.

The only ones that have not are either unique to Python (e.g.
`matrix_product` works around `numpy.dot` wanting to receive
and return C-contiguous arrays) or are only needed to aid
the Python impl. (e.g. `vec_size` is used to form a "curried"
function used to pass to the quadrature helper in
`compute_length`).