Updating Travis config.
Install:

- `brew install gcc` so that we have `gfortran`
- Gets `osx_utils.sh` (and `common_utils.sh`, which `osx_utils.sh`
  depends on) from `matthew-brett/multibuild`
- Use `install_mac_cpython` from `multibuild` to install different
  versions based on build matrix
- Make sure the newly installed python.org CPython has "clean"
  `pip` and `virtualenv`
- Install `nox` test runner
- Export `PY_BIN_DIR` so that scripts can run `${PY_BIN_DIR}/nox`

Test:

- On Python 2.7 run `cover`
- On Python 3.5 run `unit_tests`
- On Python 3.6 run `unit_tests`, `doctests` and `check_journal`

The first build [1] has **still** not begun (80 minutes in
"queued" state) but based on the `osx_utils.sh` helpers
in `matthew-brett/multibuild` (and the Travis CI docs [2])
it seems the steps above are necessary (i.e. we must install
the python.org CPython manually).

An empty `travis_journal.txt` has been added, so the first build
will intentionally fail (but it will print the diff that was
expected).

[1]: https://travis-ci.org/dhermes/bezier/builds/275722856
[2]: https://docs.travis-ci.com/user/languages/python/
     > Python builds are not available on the OS X environment.