Moving `bool` include hacks into a dedicated header file.
This was inspired (in the initial commit [1]) by a header from
`msgpack` [2] (and also some help from StackOverflow [3] about
`_MSC_VER` and Matthew Brett's pydagogue doc [4]).

The "last" modification (in [5]) diverges from the original ([2])
by using `char` instead of `int` as the `bool` alias on Windows.
This is so that we can remain true to the binary interface.

[1]: https://github.com/dhermes/bezier/commit/833643f4e8913eacbf4f04902fd00e692f0bce62
[2]: https://github.com/msgpack/msgpack-c/blob/b02c6beb4dfc28b9651ac78e583d2273aef679a7/include/msgpack/sysdep.h#L180-L195
[3]: https://stackoverflow.com/a/5850405/1068170
[4]: http://matthew-brett.github.io/pydagogue/python_msvc.html
[5]: https://github.com/dhermes/bezier/commit/6fb9da28e992bba0621304d1c8a8208827ea670e