Implementing ``all_intersections()`` in Fortran.
It currently has a (large) memory leak which I will try
to track down ASAP.

In addition, I spent a decent chunk of time tracking down a nasty SEGFAULT
caused by the ``CurveData%root`` pointer pointing to no-longer valid memory.
By using ``candidates_(even|odd)`` as workspaces and allowing
them to be over-written, the **original** roots ended up coming
from ``candidates_odd`` as it existed in the first iteration (which was
then over-written on subsequent calls). To address this, I had to add in
the hacky

    if (index_ == 1) then
       ...
    else if (is_even) then
       ...
    else
       ...
    end if

when deciding which workspace to READ from and WRITE into.

(This is yet again **more** argument against using a pointer
for ``root``, I should instead use a plain integer index
into some blessed list of candidates.)