[ci skip] Updating `bezier.Dockerfile` to also build SciPy wheel.
`scipy imports were failing with

  ImportError: unable to load extension module
  '.../.nox/unit-py-pypy/site-packages/scipy/linalg/_fblas.pypy-41.so':
  liblapack.so.3: cannot open shared object file: No such file or directory

So this required **at least** `apt install`-ing `libatlas-dev`,
`libblas-dev` and `liblapack-dev`.

I figured I might as well build a SciPy that I **know** works and put
the wheel in the wheelhouse.

Note that I install the NumPy wheel into the PyPy virtualenv
because it is required for building SciPy.