Making `all_intersections()` C-interoperable.
This is done by adding a new failure mode where the passed
in `intersections` variable may not have enough space for
all the intersections.

This is not really a large problem because the caller can
always guarantee enough space by using Bezout's theorem
and allocating the worst case (it will be
`(num_nodes_first - 1)(num_nodes_second - 1)`).