Removing `Curve._root`.
Fixes #77.

Had to do some unnecessary computation in `test_surface_surface`
to address this issue.

I will be changing the underlying data structure that defines
a curved polygon / surface-surface intersection. I plan on making
the index of the original edge part of this data structure, so
at that point we can drop the two extra calls to `.specialize`.