Introducing a `SegmentNode` helper type.
Will be used for surface-surface intersection.

Updated `get_next()` to use this and will follow up with changes
to `to_front()` that do the same.

This type is useful because it doesn't discard the useful
information that a node has now been classified and can be directly
placed on a segment of one of the six edges ({1, 2, 3} from surface 1 and
{4, 5, 6} from surface 2).

In addition, the type is "necessary" as a return value because both
`to_front()` and `get_next()` need to return a node that doesn't have
a corresponding intersection.

Will follow up later with a change that adds a call to `remove_node()`
and adds `start` (just the index) as an argument so that both `get_next()`
and `to_front()` can return a boolean indicating if we are back where
we started.