Re-factoring `get_next()` to use an `Intersection`.
After implementing `interior_combine()` and passing in the 1L-6L
intersection, I realized that when the classification **SWITCHES**
at a node, the information of the **other** parameter (e.g. ``s``
when ``t`` was used on a ``SECOND`` edge) is lost. (My code did
exactly what I told it to, but it was still the wrong thing.)

As a result, the `SegmentNode` will need to be removed.