Adding a "simple" intersection with 2 curved polygons.
This is actually to make it easier to write a few unit
tests for the (coming in the next commit) Fortran version
of `interior_combine()`.

In particular, note that all the nodes and parameters in this
intersection can be represented **exactly** in floating point
(i.e. they are rational and have a denominator that is a power
of two).