Marking `segments` and `segment_ends` as `inout`.
This way, callers can re-use the same space without having
to worry about continual re-allocation.

To do this, had to add an additional `num_intersected` to
describe the "actual" size of `segment_ends`. Once known,
the "last" value of `segment_ends` describes the "actual"
size of `segments`.

This is in preparation for exposing an ABI for `surfaces_intersect`.