Adding Cython interface for `surface_intersections()`.
This is **far** from complete, but I added it just to verify
that the ABI was working as expected. (For example, to avoid using
`malloc`, I just made `segments` a fixed size array.)

Also updated the C header and pxd declaration with the rename
of `TOO_SMALL` to `INSUFFICIENT_SPACE`.