Updating `surfaces_intersect_abi` for the "no intersections" case.
This has been failing the Mac OS X build for 15 days ([1] to [2]).

The line (1498)

    num_segments = segment_ends(num_intersected)

in `surface_intersection.f90` is nonsense when `num_intersected == 0`,
but this **never** failed on Ubuntu or Windows. On Ubuntu, I have verified
that `num_segments == 0` **always** when the "illegal" memory was accessed,
which kept from triggering `(num_segments > segments_size)`.

I also should be running tests with bounds checking on, DERP.

[1]: https://travis-ci.org/dhermes/bezier/builds/305619143
[2]: https://travis-ci.org/dhermes/bezier/builds/311996437