Making some lint-only changes.
Some are relevant (moving standard libary imports to
module level in a test and explicitly returning `None`
in a function).

The other changes reduce the number of locals based on a bug in
Pylint that adds ``__class__`` to the list of locals.

Some info about the ``__class__`` local:

https://stackoverflow.com/q/36993577/1068170

I will be filing an issue with ``pylint``.