Adding "wiggle" to Curve `locate_point()`.
Pushes values back into `[0, 1]` when they accidentally get
pushed outside `[0, 1]`.

This is "imperfect" in that `0` and `1` are still "difficult"
to detect (since the candidates will always be biased on one
side).

In practice though this won't matter, since I plan using
`locate_point()` to detect when curves are coincident
and by that point, touching endpoints should be covered
already.