Updating Python code to store nodes in columns instead of rows.
This is a follow-on to b44af8c3d590add947f905f2bc016af7272fc8e0 (the
previous commit) and like that commit, this is also a "kitchen-sink"
commit that isn't readable.

This **only** fixes the `.pyx` / `.py` source files and the
associated unit tests.

There will be more follow up commits fixing

- Functional tests
- Docs
- Doctests
- Generated Images
- Lint