Adding a `TANGENT_BOTH` classification.
This is a "robustness" change and (for now) it allows a "degenerate"
curved polygon of intersection. What will likely happen is
that the change to `surface_intersections.json` will be reverted
and code will be added to split the intersection at the
`TANGENT_BOTH` point.

See #105 for another functional test that will be added.