Returning `coincident` status from curve-curve `all_intersections()`.
This is (for now) unused in

- `surface_intersection.f90::surfaces_intersection_points()`
- `_surface_intersection.py::surface_intersections()`

but will be used in the next commit to actually correctly classify
such points as `COINCIDENT`.

Also removed the `UNCLASSIFIED` status since it will not be needed.
I originally added it because I planned on returning an integer
array of classifications from `all_intersections()` but when
I went to implement it I realized it would be more concise to
just return a boolean flag indicating if the curves are
coincident.

Also switched to using columns instead of rows in `jac_mat` in
`curve_intersection.f90::newton_refine_intersect()`. This way
a copy was not needed when passing to `evaluate_hodograph()`.
The debug flags that the Fortran unit tests are built with were
printing:

   At line 163 of file .../bezier/src/bezier/curve_intersection.f90
   Fortran runtime warning: An array temporary was created for argument 'hodograph' of procedure 'evaluate_hodograph'
   At line 167 of file .../bezier/src/bezier/curve_intersection.f90
   Fortran runtime warning: An array temporary was created for argument 'hodograph' of procedure 'evaluate_hodograph'

and now all the warnings are gone. (There were **many** such warnings
before switching from storing nodes in rows to columns.)