Adding `full_newton` for convergence near problematic intersections.
This first tries a standard Newton's method:

    p{n+1} = p{n} - DF^{-1} F

and then "upgrades" to a different function G when convergence
becomes linear. This function `G` accounts for the curves being
tangent at the point of intersection (and assumes linear
convergence is caused by the root being a double root).

This is a Python only implementation, Fortran support will be
added soon, then integrated into `from_linearized` in both
implementations.