Integrating `full_newton()` into curve-curve intersection.
This mostly "completes" the robustness fix that has been
in flight, but there is still a little bit to resolve. This
fix has introduced spurious (i.e. repeat or very inaccurate)
intersections for curve-curve functional test cases 12, 17, 43 and 44.

The fix for these spurious intersections will likely be to address
issue #109.