"Avoiding zero" when determining if an intersection is a duplicate.
Doing this by using `1 - s` rather than `s`, as is done in `full_newton()`.
This is perfectly "appropriate" since evaluating a Bezier curve requires using
both ``s`` and ``1 - s``, so both values are equally relevant. Also using a
relative error threshold rather than the more stringent "1 ULP error" check.
Used the same threshold for checking "sameness" that is used in
`full_newton()`. This is because (in the next commit) the full Newton's
method will always be used.

Also:

- Removing the `ulps_away()` helper (from Python and the ABI).
- Removing `set_similar_ulps()` / `get_similar_ulps()` (from Python and the ABI).

This fixes 2 of the 4 "broken" curve-curve functional tests and
the two that remain broken are caused by "undetected" tangencies.