Using a full Newton's method **always** in `from_linearized()`.
This will **mostly** do the same amount of work: one Newton step.
The change is to account for tangent intersections that are
"accidentally well-behaved". For example, the curve-curve
intersection functional test cases 43 and 44 have linearized
segments that intersect inside `[0, 1] x [0, 1]` even though they
"shouldn't" because of a tangency.

(An alternate approach here would be to check when the convex
hulls are tangent and do a check similar to `tangent_bbox_intersection()`.)