Fixing `NewtonDoubleRoot` in Python.
For the case where `nodes1/nodes2` was a line, it was using freed
memory as the evaluated `B1''(s)`` / ``B2''(t)`` value, which
was sometimes very large (rather than exactly zero).

See some failures:

- https://travis-ci.org/dhermes/bezier/builds/349071638
- https://circleci.com/gh/dhermes/bezier/1247

Also relaxing the requirement in `utils.almost()` that the
expected value must **differ** from the computed value.
This is because on some platforms they may be the same, while
on other platforms differ.