Adding a problematic curve-curve intersection test case.
This is from issue #32.

This breaks the assumption that we are sufficiently close to a solution
in `newton_iterate()`.

I'm still investigating but it seems that one root drifts to
the other which makes `newton_iterate()` think convergence isn't
quadratic (since the errors are large relative to the size of the
solution).