Fixing handling of problematic curve-curve intersection test case.
Case was added in 3e56fb700287489086e0bbe9a763fe149886df6f.
The fix (for now) is to raise the relative error threshold from `2^{-45}`
to `2^{-42}`.

However, it may instead make more sense to change the way that
`linear_updates` is updated. For the broken case, the relative error
is around `2^{-43}` during four iterations of the simple root case but
the values "jump around" near the solution, resulting in an error
ratio (`norm_update` / `norm_update_prev`) which exceeds `0.25` (the
ratios are `6.5e-7`, `1.72`, `0.55` and `0.77`).