Adding surface-surface intersection test case that fails unexpectedly.
This is because a tangency is "inexact" (i.e. the tangent vectors
aren't exactly orthogonal), so the point `E0(3/4) = E3(2/25)` is
is classified as `FIRST` rather than `TANGENT_FIRST`.

This can be fixed by allowing a bit of wiggle room to classify
the point as tangent (see #108) or by just "collapsing" the
two segments `(0, 0.0, 0.75)` and `(0, 0.75, 1.0)` into
`(0, 0.0, 1.0)`.