Slight modification to convex hull algorithm.
This makes sure points from the bottom of the hull can't
be used in the top. I added an example from a real-world failure
of the algorithm that tried to add a 4th point to the convex
hull when only 3 points were allocated for it.

Before the source changes, the Fortran unit test failed with

   At line 296 of file .../src/bezier/helpers.f90
   Fortran runtime error: Index '4' of dimension 2 of array 'polygon' outside of expected range (1:3)

and the Python unit test **crashed** because bounds checking is off (by
default) in the Python unit tests (see #89):

   *** Error in `.../bin/python3.6': free(): invalid next size (fast): 0x0000000002f15d40 ***
   ======= Backtrace: =========
   /lib/x86_64-linux-gnu/libc.so.6(+0x777e5)[0x7ff8a94247e5]
   /lib/x86_64-linux-gnu/libc.so.6(+0x8037a)[0x7ff8a942d37a]
   /lib/x86_64-linux-gnu/libc.so.6(cfree+0x4c)[0x7ff8a943153c]
   .../numpy/core/multiarray.cpython-36m-x86_64-linux-gnu.so(+0x1eb7c)[0x7ff8a423eb7c]