Lowering wiggle and Newton thresholds.
Adding 4 surface-surface functional test cases (and associated
curve-curve functional tests for 3 of them) that showcase
this failure.

In three of the cases, the `NEWTON_ERROR_RATIO` was too strict,
causing `newton_iterate()` to exit with `NewtonSimpleRoot()`
thinking it hadn't converged. In all cases, the normalized
tangent vectors (i.e. T(s) = B(s) / ||B(s)||) were near parallel, but
not parallel. (For comparison, the **actually** tangent test cases
have cross product `T1(s) x T2(t)` on the order of `10^{-6}` or
smaller.)

- 48Q-14L (49) (also curve-curve case 51 between 73-74). The
  normalized tangent vectors had a cross-product of `-0.0183174`.
  During ``newton_iterate()`, the relative error ratio stalls out near
  `2^{-40.9}` which was too large when the threshold was set at
  `2^{-42}`.
- 49Q-15L (50) (also curve-curve case 52 between 75-76). The
  normalized tangent vectors had a cross-product of `-0.0269569`.
  During ``newton_iterate()`, the relative error ratio stalls out near
  `2^{-38.6}` which was too large when the threshold was set at
  `2^{-42}`.
- 3C-4C (51) (also curve-curve case 53 between 77-78). The
  normalized tangent vectors had a cross-product of `-0.00118459`.
  During ``newton_iterate()`, the relative error ratio stalls out near
  `2^{-35}` which was too large when the threshold was set at
  `2^{-42}`. However, one of the iterates does have a relative error of
  `2^{-36.4}` and thats what we base the new `NEWTON_ERROR_RATIO` on.

In the fourth case (50Q-51Q, case 52), a spurious intersection near a corner
caused `basic_interior_combine()` to fail to converge. Both
`I1: B1(0.0) = B3(0.6470588235294181)` and
`I2: B0(0.9999999999999566) = B3(0.647058823529457)` are found as intersections
in addition to a corner `I3: B3(1.0) = B4(0.0)` and the other intersection
`I4: B0(0.6378830116519687) == B4(0.0040913069056350065)`.
(When computing the intersection in high precision, the near-corner
intersection actually occurs at
`B1(2.4524788197948127e-15) == B3(0.6470588235294156)`.)
The issue occurs because the "interior combine" process gets stuck in an
edge loop:

   I1 -> [I2 -> I3 -> I4] -> [I2 -> I3 -> I4] -> [I2 -> I3 -> I4] -> ...

that never terminates. By making `wiggle_interval()` slightly more tolerant,
the intersection `I2` gets `B1(0.0)` and is discarded as a duplicate. In this
case, the intersection polygon follows:

   Edge 0: I1 -> I3 (along B3)
   Edge 1: I3 -> I4 (along B4)
   Edge 2: I4 -> I1 (along B0)

This is still "wrong" but is the best we can expect since the true intersection
happens so near a corner (`B1(2.4524788197948127e-15)`) that we can't detect
it (the parameter value we detect that gets "wiggled" to `0.0` is close, but
not that close: `8.215650382226158e-15`). The true intersection is a curved
quadrilateral (using another corner `I5: B0(1.0) = B1(0.0)` and the "true"
intersection `I6: B1(2.4524788197948127e-15) == B3(0.6470588235294156)`):

   Edge 0: I3 -> I4 (along B4)
   Edge 1: I4 -> I5 (along B0)
   Edge 2: I5 -> I6 (along B1)
   Edge 3: I6 -> I3 (along B3)

---------

Post-Script: I don't think it's feasible to try to accommodate intersections
of the second type, since parameter values so close to zero are very
difficult to deal with. It **may** be possible to do by re-parameterizing
to [-0.5, 0.5], but by doing this, the nodes will be changed inherently
due to roundoff.

For the first type, I think we can probably take `NEWTON_ERROR_RATIO` back
to be more strict at `2^{-45}`, but just need to add a secondary check
that the "normalized" tangent vectors aren't parallel. (I think just
`sqrt(machine precision)` would be a fine relative error.)