Adding full documentation for ABI.
Fixes #63. I'd like to add more examples but this is all
for now (the only example is in `status.rst`).

In addition I'd like to revisit this and make it autogenerated from
prose written either in the header files or in the Fortran files.
For now, this was **all** written by hand.

Along the way, a few small updates were made to comments in
some `.f90` files.