Putting quotes are `SET PATH=...` in `.appveyor.yml`.
This matches the demo [1]. The quotes are there so the
double slash `\\` is escaped into a single slash and
to enable values that might have spaces in the environment
variables.

For some reason, builds on AppVeyor had started to fail [2]
due to `C:\Python27-x64\\Scripts` being set on the PATH
rather than `C:\Python27-x64\Scripts`.

[1]: https://github.com/ogrisel/python-appveyor-demo/blob/f54ec3593bcea682098a59b560c1850c19746e10/appveyor.yml#L103
[2]: https://ci.appveyor.com/project/dhermes/bezier/build/1.0.955.master/job/u5gbct3dd1d3rrft