Making command journal deterministic.
Does so by limiting `NPY_NUM_BUILD_JOBS` [1] to 1. For Fortan 77 files,
the default compile step [2] uses `N` jobs where `N = get_num_build_jobs()`,
and then builds in parallel. Since the execution of these subprocesses is
up to the OS, the order is not deterministic (for some reason there were
never issues on Ubuntu 16.04). I initially thought it was dictionary
order when the sources were parsed [3] but then I realized that dictionary
insertion order is preserved in Python 3.7.

[1]: https://github.com/numpy/numpy/blob/v1.15.0/numpy/distutils/misc_util.py#L100
[2]: https://github.com/numpy/numpy/blob/v1.15.0/numpy/distutils/ccompiler.py#L334
[3]: https://github.com/python/cpython/blob/v3.7.0/Lib/distutils/ccompiler.py#L343