Removing `bool` from `nitpick_ignore`.
See comment in `docs/conf.py`. This is because `_Bool` is part of
the language (as of C99) but `bool` is not. So I added `bool` to the
set of accepted "stopwords" for the C domain.

I also added Sphinx 1.7.0 as the minimum to ensure I'm not using
features that are too new.