Ignoring `_pypy_speedup.c` when regenerating.
This is also a temporary hack. See failure:
https://circleci.com/gh/dhermes/bezier/1466

The near term fix will be to generate both `_speedup.c` and
`_pypy_speedup.c` in the same script.

The long term fix will hopefully be a removal of `_pypy_speedup.c`
once PyPy releases a new version that supports latest NumPy.