Making `gcc` snippets in `docs/abi/` more copy-paste friendly.
By using environment variables for the `-L` (lib) and `-I` (include)
arguments, the entire `gcc` invocation can be copy-pasted without
modification so long as `INCLUDE_DIR` and `LIB_DIR` are set correctly.

This **does not** solve the issue of the `$ ` and `> ` prefix on each line,
which is another annoying impediment to copy-paste-ability.