Converting console codeblocks to doctest for Linux sections of binary extension.
This is **almost certainly** more trouble than it's worth. To get this
working "correctly" the `doctest` has to be able to install a wheel
created in the `manylinux` container (i.e. with an "old enough" `libc`).
This is bad enough on a host OS during local dev, but on CircleCI, this
requires docker-in-docker (which luckily isn't too challenging since
`/var/run/docker.sock` can be shared).

However, what **ultimately** may shoot this down is the fact that
the hash on `libbezier-*.so` produced by `auditwheel` (which is just
the first 8 characters of the SHA256 checksum of the `.so` file,
references [1] and [2]) seems to be non-deterministic.

[1]: https://github.com/pypa/auditwheel/blob/d5cd0f670eecf9a94be67911e18d3c2dea8ffc55/auditwheel/repair.py#L127
[2]: https://github.com/pypa/auditwheel/blob/d5cd0f670eecf9a94be67911e18d3c2dea8ffc55/auditwheel/hashfile.py#L4-L11