Removing normalization from `clipping`.
Also removes some other unneeded scaling. The primary motivation
for me here was to make it more possible to get an **exact** answer
(of `0.875`) out of my handcrafted example. However, when I actually
dug into where the roundoff was occurring, I realized a lot of the
scaling / normalizing wasn't necessary and was introducing the
roundoff cost as along with extra flops.