Stop using NORMALIZE_WHITESPACE for Windows.
The only thing that needed to be normalized was line endings (because
the output will have Windows line endings but the content it is
expected to match is authored with UNIX line endings). However the
use of NORMALIZE_WHITESPACE means we also ignored other whitespace
differences which we didn't intend to.